eval/logical_chain_strength/requirements.txt
# For LLM interaction
langchain>=0.1.0
langchain-openai>=0.3.18
python-dotenv

# For LaTeX parsing (if needed, can be done with regex initially)
# Add other specific libraries here if required
